# swaggerui/swagger.yml
swagger: "2.0"
info:
  title: Book Library
  description: |
    This is a sample API that describes the structure of our Book-Library-Server
  version: 1.0.0
  license:
    name: Apache-2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.html'
...
paths:
  /v1/users/create:
    post:
      tags:
        - users
      operationId: NewUser
      summary: create new users
      security:
        - sessionCookie: [ ]
      description: |
        Simple create interface used to create a new user in the databse
      parameters:
        #header
        - $ref: "#/components/headers/X-Request-ID"
      requestBody:
        description: Create user requestBody
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/newUser"
        required: true
      x-codegen-request-body-name: NewUser
      responses:
        '202':
          $ref: "#/components/responses/202_NewUser"
          description: sucessful action
        '401':
          $ref: "#/components/responses/401_Unauthorized"
          description: user not authorized
        '400':
          $ref: "#/components/responses/400_InvalidID"
          description: Invalid ID value
...
  /v1/users/{user-id}:
    get:
      tags:
        - users
      summary: Finds user by user-id
      description: |
        Simple getter interface used to get an existing user by its ID
      operationId: Retreive
      security:
        - sessionCookie: [ ]
      parameters:
        #header
        - $ref: "#/components/headers/X-Request-ID"
        #path
        - $ref: "#/components/parameters/user-id"
      responses:
        '200':
          description: successful operation
          $ref: '#/components/responses/200_UserInfo'
        '401':
          $ref: "#/components/responses/401_Unauthorized"
          description: user not authorized
        '404':
          $ref: "#/components/responses/404_NotFound"
          description: Not found
...
  /v1/books/create:
    post:
      tags:
        - books
      operationId: NewBook
      summary: create new books
      security:
        - sessionCookie: [ ]
      description: |
        Simple create interface used to create a new book in the databse
      parameters:
        #header
        - $ref: "#/components/headers/X-Request-ID"
      requestBody:
        description: Create user requestBody
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/newBook"
        required: true
      x-codegen-request-body-name: NewBook
      responses:
        '202':
          $ref: "#/components/responses/202_NewBook"
          description: sucessful action
        '401':
          $ref: "#/components/responses/401_Unauthorized"
          description: entity not authorized
        '400':
          $ref: "#/components/responses/400_InvalidID"
          description: Invalid ID value
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name:
      description: >
        keeps the connection between the UI and the server and it sets in cookies in the header-response in the browser and
        is used to verified the identity of the sender each time he calls the server.
  headers:
    X-Request-ID:
      name: X-Request-ID
      required: true
      description: |
        In Golang a Unique ID is used in the context-structure to make sure
        each time a request get through an API it can be identified properly and be canceled any time.
      example: "99391c7e-ad88-49ec-a2ad-99ddcb1f7721"
      schema:
        type: string
        format: uuid
...
  parameters:
    SESSION-COOKIE:
      name:
      required: true
      in: header
      schema:
        type: string
        example: =upUySJYvsUjRuALYqd0Lr82HGz4RZNb_Mxu9WVOQHJGXa71z
    user-id:
      name: user_id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    202_NewUser:
      description: |
        The new created user. Only admin user can create new user and seed it in the database.
      headers:
        X-Request-ID:
          $ref: "#/components/parameters/X-Request-ID"
        SESSION-COOKIE:
          $ref: "#/components/parameters/SESSION-COOKIE"
      content:
        application/json:
          schema:
            type: object
            properties:
              newUser:
                $ref: "#/components/schemas/newUser"
...
    200_UserInfo:
      description: |
        This object returns an available users. Only an admin or the user itself can proceed this action.
      headers:
        X-Request-ID:
          $ref: "#/components/parameters/X-Request-ID"
        SESSION-COOKIE:
          $ref: "#/components/parameters/SESSION-COOKIE"
      content:
        application/json:
          schema:
            type: object
            properties:
              userInfo:
                $ref: "#/components/schemas/userInfo"
...
    401_Unauthorized:
      description: |
        Provided credentials are wrong, not provided or invalid or user has not required roles for executing this action.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/X-Request-ID"
        SESSION-COOKIE:
          $ref: "#/components/parameters/SESSION-COOKIE"
      content:
        application/json:
          schema:
            type: object
            properties:
              errorResponse:
                $ref: "#/components/schemas/errorResponse"
    404_NotFound:
      description: The required resources was not found
      headers:
        X-Request-ID:
          $ref: "#/components/headers/X-Request-ID"
      content:
        application/json:
          schema:
            type: object
            items:
              $ref: "#/components/schemas/errorResponse"

    #----------BOOKS-----------
    202_NewBook:
      description: |
        The new created book. Only admin user can create new book and seed it in the database.
      headers:
        X-Request-ID:
          $ref: "#/components/parameters/X-Request-ID"
        SESSION-COOKIE:
          $ref: "#/components/parameters/SESSION-COOKIE"
      content:
        application/json:
          schema:
            type: object
            properties:
              newBook:
                $ref: "#/components/schemas/newBook"
...
  schemas:
...
    newUser:
      title: New User
      type: object
      required:
        - name
        - email
        - password
        - passwordConfirm
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        passwordConfirm:
          type: string
...
    userInfo:
      title: User Infos
      type: object
      required:
        - email
        - name
      properties:
        name:
          type: string
        email:
          type: string
...
    newBook:
      title: New Book
      type: object
      required:
        - title
        - isbn
        - category
        - describtion
        - authors
      properties:
        title:
          type: string
        isbn:
          type: string
        category:
          type: string
        description:
          type: string
        authors:
          type: array
          items:
            type: string
...
    errorResponse:
      title: ErrorResponse
      type: object
      properties:
        userMessage:
          type: array
          items:
            $ref: "#/components/schemas/userMessage"
    userMessage:
      title: userMessage
      type: object
      required:
        - category
        - code
      properties:
        category:
          type: string
          enum:
            - ERROR
            - WARNING
            - INFO
        code:
          type: string
        path:
          type: string
        text:
          type: string

#    Reset-Cookies:
#      name: Reset-Cookie
#      description: |
#        Delete all relevant cookies
#      schema:
#        type: string
#        example: |
#          "RedirectSession=null; Path=/; Max-Age=0, Session=null; Path=/; Max-Age=0"